requirements.txt
python-telegram-bot==21.4
python-decouple==3.8
cohere==5.8.1
